requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
peewee==3.16.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
cryptography>=41.0.0
requests>=2.31.0
